Fixed single fold training

+++ omr/imageoperations/music_line_operations.py
@@ -207,12 +207,12 @@ class ImageExtractDewarpedStaffLineImages(ImageOperation):
             if top_to_add < 0:
                 out.image = out.image[-top_to_add:, :]
             elif top_to_add > 0:
-                out.image = np.vstack((np.zeros((top_to_add, width), dtype=np.uint8), out.image))
+                out.image = np.vstack((np.zeros((top_to_add, width), dtype=out.image.dtype), out.image))
 
             if bot_to_add < 0:
                 out.image = out.image[:bot_to_add, :]
             elif bot_to_add > 0:
-                out.image = np.vstack((out.image, np.zeros((bot_to_add, width), dtype=np.uint8)))
+                out.image = np.vstack((out.image, np.zeros((bot_to_add, width), dtype=out.image.dtype)))
 
             if out.image.shape[0] != pre_out_height:
                 raise Exception('Shape mismatch: {} != {}'.format(out.image.shape[0], pre_out_height))
